+++ bootique-cayenne-test/src/main/java/io/bootique/cayenne/test/CayenneTableManager.java
@@ -85,6 +85,5 @@ public class CayenneTableManager {
 
             return CayenneModelUtils.createTableModel(channel, entity);
         });
-
     }
 }