+++ bootique-cayenne-test/src/main/java/io/bootique/cayenne/test/CayenneTestDataManagerBuilder.java
@@ -6,18 +6,21 @@ import org.apache.cayenne.exp.Property;
 import org.apache.cayenne.map.DbEntity;
 import org.apache.cayenne.map.EntityResolver;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashSet;
 import java.util.Objects;
+import java.util.Set;
 
 /**
  * @since 0.24
  */
 public class CayenneTestDataManagerBuilder {
+
     private BQRuntime runtime;
     private boolean deleteData;
     private boolean refreshCayenneCaches;
-    private List<DbEntity> entities;
+    private Set<DbEntity> dbEntities;
+    private Set<DbEntity> dbEntityGraphRoots;
+
     private EntityResolver resolver;
 
     CayenneTestDataManagerBuilder(BQRuntime runtime) {
@@ -27,16 +30,20 @@ public class CayenneTestDataManagerBuilder {
 
         ServerRuntime serverRuntime = runtime.getInstance(ServerRuntime.class);
         this.resolver = serverRuntime.getDataDomain().getEntityResolver();
-        this.entities = new ArrayList<>();
+        this.dbEntities = new HashSet<>();
+        this.dbEntityGraphRoots = new HashSet<>();
     }
 
     public CayenneTestDataManager build() {
+
+        Set<DbEntity> mergedEntities = mergeEntities();
+
         return new CayenneTestDataManager(
                 runtime.getInstance(ServerRuntime.class).getDataDomain(),
                 runtime.getInstance(CayenneTableManager.class),
                 deleteData,
                 refreshCayenneCaches,
-                CayenneModelUtils.tablesInInsertOrder(runtime, entities));
+                CayenneModelUtils.tablesInInsertOrder(runtime, mergedEntities));
     }
 
     public CayenneTestDataManagerBuilder doNotDeleteData() {
@@ -57,34 +64,59 @@ public class CayenneTestDataManagerBuilder {
 
         Objects.requireNonNull(entityTypes);
         for (Class<?> type : entityTypes) {
-            entities.add(CayenneModelUtils.getDbEntity(resolver, type));
+            dbEntities.add(CayenneModelUtils.getDbEntity(resolver, type));
+        }
+
+        return this;
+    }
+
+    public CayenneTestDataManagerBuilder entitiesAndDependencies(Class<?>... entityTypes) {
+
+        Objects.requireNonNull(entityTypes);
+        for (Class<?> type : entityTypes) {
+            dbEntityGraphRoots.add(CayenneModelUtils.getDbEntity(resolver, type));
         }
 
         return this;
     }
 
     public CayenneTestDataManagerBuilder relatedTables(Class<?> entityType, Property<?> relationship) {
-        CayenneModelUtils.getRelatedDbEntities(resolver, entityType, relationship).forEach(entities::add);
+        CayenneModelUtils.getRelatedDbEntities(resolver, entityType, relationship).forEach(dbEntities::add);
         return this;
     }
 
-    public CayenneTestDataManagerBuilder tableName(String tableName) {
-        return tableNames(tableName);
+    public CayenneTestDataManagerBuilder table(String tableName) {
+        return tables(tableName);
     }
 
-    public CayenneTestDataManagerBuilder tableNames(String... tableNames) {
+    public CayenneTestDataManagerBuilder tables(String... tableNames) {
 
         Objects.requireNonNull(tableNames);
         for (String table : tableNames) {
-
-            DbEntity entity = resolver.getDbEntity(table);
-            if (entity == null) {
-                throw new IllegalArgumentException("Not a Cayenne-managed table: " + table);
-            }
-
-            entities.add(entity);
+            dbEntities.add(CayenneModelUtils.getDbEntity(resolver, table));
         }
 
         return this;
     }
+
+    public CayenneTestDataManagerBuilder tablesAndDependencies(String... tableNames) {
+
+        Objects.requireNonNull(tableNames);
+        for (String table : tableNames) {
+            dbEntityGraphRoots.add(CayenneModelUtils.getDbEntity(resolver, table));
+        }
+
+        return this;
+    }
+
+    private Set<DbEntity> mergeEntities() {
+        if(dbEntityGraphRoots.isEmpty()) {
+            return dbEntities;
+        }
+
+        Set<DbEntity> merged  = ModelDependencyResolver.resolve(dbEntityGraphRoots);
+        merged.addAll(dbEntities);
+        return merged;
+    }
+
 }